+++ mud-common-client/src/main/java/com/jpinfo/mudengine/common/being/BeingClass.java
@@ -26,9 +26,9 @@ public class BeingClass implements Serializable {
 
 	private Integer weightCapacity;
 
-	private Map<String, Long> attrs;
+	private Map<String, Integer> attrs;
 
-	private Map<String, Long> skills;
+	private Map<String, Integer> skills;
 	
 	
 	public BeingClass() {
