+++ mud-common-client/src/main/java/com/jpinfo/mudengine/common/being/BeingSkillModifier.java
@@ -9,6 +9,6 @@ public class BeingSkillModifier {
 	private String originCode;
 	private String originType;
 
-	private double offset;
+	private float offset;
 	private Integer endTurn;
 }
